refactor: support logic
