+++ tests/support/data/licence.js
