+++ tests/support/scenarios/core-licence.js
